Rename raw queries feature to be explicit about SQL
